+++ stfblender/stf_resources/stf/stf_prefab/draw_prefab_ui.py
@@ -4,7 +4,7 @@ import bpy
 def draw_prefab_ui(layout: bpy.types.UILayout, context: bpy.types.Context, collection: bpy.types.Collection, operator_set_as_root: str, operator_set_id: str, operator_add_component: str, operator_remove_component: str, operator_edit_component_id: str):
 	from .....stfblender_common.helpers import draw_slot_link_warning
 	from .....stfblender_common.slot_link import get_slot_link_data_model_version
-	from ....ui.operators.stf_meta import draw_meta_editor
+	from ....ui.stf_meta import draw_meta_editor
 	from ....ui import draw_components_ui, draw_stf_id_ui
 
 	draw_slot_link_warning(layout)