+++ stfblender/stf_resources/stfexp/stfexp_animation_blendtree.py
@@ -1,7 +1,7 @@
 import bpy
 from typing import Any
 
-from ....stfblender_common import STF_ImportContext, STF_ExportContext, STF_TaskSteps, STF_Category, STFReport, STF_DataResourceBase, STF_Handler_Data, STF_Handler_ComponentHolder, STF_Data_Ref, add_resource, export_data_resource_base, get_components_from_data_resource, import_data_resource_base
+from ....stfblender_common import STF_ImportContext, STF_ExportContext, STF_TaskSteps, STF_Category, STFReport, STF_NonNativeResourceBase, STF_Handler_NonNative, STF_Handler_ComponentHolder, STF_NonNativeResource_Ref, add_nonnative_resource, export_nonnative_resource_base, get_components_from_nonnative_resource, import_nonnative_resource_base
 from ....stfblender_common.helpers import draw_list, poll_valid_animations
 
 
@@ -9,7 +9,7 @@ class BlendtreeAnimationMapping(bpy.types.PropertyGroup):
 	position: bpy.props.FloatVectorProperty(name="Position", size=2, default=(0, 0), soft_min=-1, soft_max=1, options=set(), subtype="XYZ")
 	animation: bpy.props.PointerProperty(name="Animation", type=bpy.types.Action, options=set(), poll=poll_valid_animations)
 
-class STFEXP_Animation_Blendtree(STF_DataResourceBase):
+class STFEXP_Animation_Blendtree(STF_NonNativeResourceBase):
 	type: bpy.props.EnumProperty(name="Type", items=(("1d", "1D", ""), ("2d", "2D", "")))
 	animations: bpy.props.CollectionProperty(type=BlendtreeAnimationMapping, options=set())
 
@@ -31,7 +31,7 @@ def _draw_func_animation1d(layout: bpy.types.UILayout, element: Any) -> bpy.type
 	return row
 
 
-class Handler_STFEXP_Animation_Blendtree(STF_Handler_Data, STF_Handler_ComponentHolder):
+class Handler_STFEXP_Animation_Blendtree(STF_Handler_NonNative, STF_Handler_ComponentHolder):
 	"""Define a blendtree for animations"""
 	stf_type = "stfexp.animation_blendtree"
 	stf_category = STF_Category.DATA
@@ -39,7 +39,7 @@ class Handler_STFEXP_Animation_Blendtree(STF_Handler_Data, STF_Handler_Component
 	blender_property_name = "stfexp.animation_blendtree"
 
 	@classmethod
-	def draw(cls, layout: bpy.types.UILayout, context: bpy.types.Context, resource_ref: STF_Data_Ref, context_resource: Any, resource: STFEXP_Animation_Blendtree):
+	def draw(cls, layout: bpy.types.UILayout, context: bpy.types.Context, resource_ref: STF_NonNativeResource_Ref, context_resource: Any, resource: STFEXP_Animation_Blendtree):
 		layout.prop(resource, "type")
 
 		if(resource.type == "2d"):
@@ -49,8 +49,8 @@ class Handler_STFEXP_Animation_Blendtree(STF_Handler_Data, STF_Handler_Component
 
 	@classmethod
 	def import_resource(cls, context: STF_ImportContext, json_resource: dict, stf_id: str, context_resource: Any | None) -> Any | STFReport:
-		resource_ref, resource = add_resource(context.get_root_collection(), cls.blender_property_name, stf_id, cls.stf_type)
-		import_data_resource_base(resource, json_resource)
+		resource_ref, resource = add_nonnative_resource(context.get_root_collection(), cls.blender_property_name, stf_id, cls.stf_type)
+		import_nonnative_resource_base(resource, json_resource)
 		resource.type = json_resource["blendtree_type"]
 
 		def _handle():
@@ -68,7 +68,7 @@ class Handler_STFEXP_Animation_Blendtree(STF_Handler_Data, STF_Handler_Component
 
 	@classmethod
 	def export_resource(cls, context: STF_ExportContext, blender_resource: STFEXP_Animation_Blendtree, context_resource: Any | None) -> tuple[dict, str] | STFReport:
-		ret = export_data_resource_base(context, cls.stf_type, blender_resource)
+		ret = export_nonnative_resource_base(context, cls.stf_type, blender_resource)
 		ret["blendtree_type"] = blender_resource.type
 
 		def _handle():
@@ -86,7 +86,7 @@ class Handler_STFEXP_Animation_Blendtree(STF_Handler_Data, STF_Handler_Component
 
 		return ret, blender_resource.stf_id
 
-	get_components = get_components_from_data_resource
+	get_components = get_components_from_nonnative_resource
 
 
 def register():