+++ changelog.md
@@ -1,8 +1,10 @@
 # Changelog
 
 ## v0.1.12
-* Select how to export each Blender native resource
+* Select how to export each Blender native resource.
 * Removed `boilerplate_(un)register` in common. All Blender resources that are relevant for STF get the appropriate stf info registered by default, even if not supported by default.
+* Renamed stf_data_resource to non-native resource for better clarity.
+* Refactored blender_grr code.
 
 ## v0.1.11
 * Code cleanup & improved type annotations in common.